requirements.txt
easyocr
python-dotenv
kafka-python